Stress test with a layer of 100_000 random points
Added a new layer with 100_000 random points. The points are generated
randomly using the `randomPoints` function from the `turf` library.

Adding 100_000 points to the screen is very slow and resource intensive,
so I added an event listener to the map that listens for the `moveend`
event.
When this event is fired, the map will filter the points that are on
the screen and only display those, using the `filter` attribute of the
option argument of geoJSON function.

For now, we remove the layer and recreate it every time the map is
moved. We could perhaps mutate the layer for better performances.